+++ rust/rfprop/src/sigserve.h
@@ -9,8 +9,8 @@ struct TerrainProfile;
 
 int init(const char *sdf_path, bool debug);
 Report handle_args(int argc, char *argv[]);
-TerrainProfile terrain_profile(double tx_lat, double tx_lon, double tx_antenna_alt_m, double rx_lat, double rx_lon,
-                               double rx_antenna_alt_m, double freq_hz, bool normalize, bool metric);
+TerrainProfile terrain_profile(float tx_lat, float tx_lon, float tx_antenna_alt_m, float rx_lat, float rx_lon,
+                               float rx_antenna_alt_m, float freq_hz, bool normalize, bool metric);
 
 }  // namespace sigserve_wrapper
