+++ src/models/fspl.cc
@@ -26,8 +26,8 @@ static __inline float _20log10f(float x) {
     return (8.685889f * logf(x));
 }
 
-double FSPLpathLoss(float f, float d, bool bBlock) {
-    double dLoss = 0.0;
+float FSPLpathLoss(float f, float d, bool bBlock) {
+    float dLoss = 0.0;
 
     if (bBlock) {
         // No line of sight
