+++ src/signal-server.hh
@@ -5,29 +5,29 @@
 
 #include "common.hh"
 
-int ReduceAngle(double angle);
-double LonDiff(double lon1, double lon2);
-void * dec2dms(double decimal, char * string);
-int PutMask(struct output * out, double lat, double lon, int value);
-int OrMask(struct output * out, double lat, double lon, int value);
-int GetMask(struct output * out, double lat, double lon);
-void PutSignal(struct output * out, double lat, double lon, unsigned char signal);
-unsigned char GetSignal(struct output * out, double lat, double lon);
-double GetElevation(site const & location);
-int AddElevation(double lat, double lon, double height, int size);
-double Distance(site const & site1, site const & site2);
-double Azimuth(site const & source, site const & destination);
-double ElevationAngle(site const & source, site const & destination);
-double ElevationAngle2(Path const & path,
+int ReduceAngle(float angle);
+float LonDiff(float lon1, float lon2);
+void * dec2dms(float decimal, char * string);
+int PutMask(struct output * out, float lat, float lon, int value);
+int OrMask(struct output * out, float lat, float lon, int value);
+int GetMask(struct output * out, float lat, float lon);
+void PutSignal(struct output * out, float lat, float lon, unsigned char signal);
+unsigned char GetSignal(struct output * out, float lat, float lon);
+float GetElevation(site const & location);
+int AddElevation(float lat, float lon, float height, int size);
+float Distance(site const & site1, site const & site2);
+float Azimuth(site const & source, site const & destination);
+float ElevationAngle(site const & source, site const & destination);
+float ElevationAngle2(Path const & path,
                        site const & source,
                        site const & destination,
-                       double er,
+                       float er,
                        LR const & lr);
-double ReadBearing(char * input);
+float ReadBearing(char * input);
 void ObstructionAnalysis(Path const & path,
                          site const & xmtr,
                          site const & rcvr,
-                         double f,
+                         float f,
                          FILE * outfile,
                          LR const & lr);
 void resize_elev(struct output & out);
